Fix S3 paths for WP subdir installs w/upload_path
Where WP is installed an a subfolder and a custom upload_path is set (as
is the case for multisite installations) the dynamic prefix for the S3
file name includes the entire WP directory name on the server.

This is ugly and exposes internal details that should probably not be
public.

The fix avoids using ABSPATH (which has the subdirectory name in it) and
instead uses a guess at the actual install directory.

If the content directory is installed inside the WP subdirectory (an
unusual scenario), the current problem will remain.  If WP provides a
better method for determining the root directory then this could also be
resolved.