Fix import of existing items
If item exists in the as3cf_items table `wpdb::update` returns int 0, which is incorrectly regarded as error. Only boolean false indicates error, so I added a strict type check. In fact it is the behavior, that was present in the 2.3.1 version and the bug was introduced in the following commit:

https://github.com/deliciousbrains/wp-amazon-s3-and-cloudfront/commit/0142a3b9bd32c9d16529ab0d6f0bb0eeaf65715d#diff-319ee688c474fcd2a3bde8b6e7ee48ffL381